ALGatorWeb/requirements_old.txt
Django==3.2
django_pygments
djangorestframework
requests
mysqlclient
werkzeug
pip-system-certs
django-extensions
pyOpenSSL
djangorestframework-simplejwt
djangorestframework-simplejwt[crypto]
